Rename ChannelRow
