+++ nextjs/components/Pages/Channel/Content/Row/index.module.scss
